codedReturn: add some self-tests
Can't import from `11_...` so I split the actual routines out into
`return_tools.py` instead.  Maybe the names can be improved...
Also, add a front-end function that opens files so that we can pass
`StringIO` objects to the actual function.  This is useful in testing
because we can pass strings of csv content.