canvas csv: don't assume Section column is integer
Use "object" dtype for all columns.  Otherwise, integer columns
become floats in the presense of missing data.  If we need special
columns to be int, we can cast them to int when needed.

One place this might be needed is the merge.  But loading the other
grade data as object works too so do that.  Probably all this code
works with general string student "numbers" (but not tested).

Take home message: pandas dtypes are for very careful datasets with
very homogeneous columns, neither of which is true of Canvas data.