text comments starting with "tex" will now be auto-latexed
If a comment starts with "tex" (upper or lower case) then
the system will automatically latex it when pasting it.
This also happens when a text item starts with "tex" and
the user ends the editing with either shift-enter or ctrl-enter.
If the user ends the edit with esc, then latex is not run.