refactor code to allow identifying with fewer QR codes
If we have one or two QR codes, we can still proceed with identification.
We warn the user (b/c if it happens a lot, its b/c of a poor scan/print).
In principle, one is riskier than two: we keep an "riskiness" value to
indicate this.  Some future UI might use these values, possibly with
other info to decide whether to bother a human being or not.

For now, all these cases just become failures anyway because the
orientation cannot generally be known in (some of) such cases.