Reassemble each page as sheet of letter paper
We make each page landscape or portrait by specifying the "page"
size as "792 x 612 pts" or "612 x 792 pts".  This is different from
rotating the image because we want the landscape pages to appear
sideways in PDF viewers, but still print properly.

Fixes #201.

Imagemagick is a bit finicky about all this.  Its not clear to me
how to center the image on the page without messing around with
density or resize.  For now, leave a TODO, and let it attach to
one edge: good enough for now.