put "plom" in the tmpdir name
Plom puts its temporary files in a system-specific temp dir.  This is good
but hard to see what is plom if you need to debug.  This change prefixes
plom temp dirs with `plom_` so they are easy to see.  That is we get
`/tmp/plom_bunbhzi6/` instead of `/tmp/tmpq8hqnct_`.