wait for thread instead of watching for file to exist
Don't know if writes to disk will be atomic and given that its writing
from another thread, this is probably safer.  I can see problems with
future improvments to the downloading which makes the coupling weaker
so leave a TODO about how to improve this more during future Downloader
rewording.