Marker: don't move the view forward on download
We will move there when we want to, thank you very much.  This means
if someone looks at the Marker window while annotating, it will show
the paper they are currently marking, not the next one.  This also means
if we cancel, we're still on the correct entry on return from Annotator.