Marker: refactor code to use updateProgress
But you can pass in values if you know them to avoid an extra round trip
if we need to ask the server.  This is better than modifying the idProgressBar
widget directly b/c we want to muck around with its text field in the (0,0)
case.  A side effect is we can relax the "greater than" checks.