compact->wide: match modeLayout alignment
I don't understand why these are here, but that's how it is in the
autogen file, so restore it to that form.  As an aside, I really dislike
this mix-and-match of "oh our UI is autogen from qtCreator stuff" and
"here are some manually mucking around".  I'm ok with either; not both.
But this little commit cannot fix all that.