respect the skip img extraction flag
By default, we do some sanity checks and then extract raster/bitmap
images when it seems sane to do so.  As this process is not yet perfect
its too dangerous to do this for student-provided input, so in HW-mode
we skip-by-defaut.  For standard admin scans, we leave the default as
extract-when-possible.