questions always list in low-level functions, fix allbyq
allbyq should not try to process the underscore q values.  In addition,
the logic is that q is either "_" or if its not that it must match the
contents of the questions list.  And in low-level functions that is
always a list, not a scalar, as is document in the docstrings.

Throw exceptions when unexpected things happen: otherwise users might
miss uploading something which is very scary.  Print-and-continue is ok
only in the "already uploaded -- skipping" case.