refactor code to publically access spec
The spec object itself is a central place to decide what info could
reasonably be considered "public".  For now, we strike the private seed,
same as the server used to do.  Now the server can just pass on whatever
it gets from this method call.

Also, clean up a TODO for a simplier calling sequence.