Depend on file-magic, somewhat loosely
Various of our contrib scripts use `import magic`.  It would be nice to
have this inside the Docker image or virtualenvs.

Unfortunately, there are two packages, which complicates packaging.  For
now we are not adding it to `setup.py`.  That will have to change if
this becomes a "core" feature.

This is all documented in [1].

[1] https://gitlab.com/plom/plom/-/issues/1570