Server: add command line args to control logging
One can turn of the stderr logging (`--no-logconsole`) and/or specify a
logfile.  If the logfile is omitted with stick with own current default
of `plomserver-<timestamp>.log`.

PlomServer and subclasses gain a new `demo.logfile` property which
returns a `pathlib.Path` to the logfile of the underlying server, which
is currently hardcoded to be `<basedir>/server.log`.