latex: make dvipng picky and return errors to client
dvipng could silently make bad/blank/incomplete pngs, for example if you
use pstricks.  Pass it `--picky` which fixes this in the sense that it
now errors out.  But also return errors in the API call so clients could
in principle show the error to their users.

Fixes #1523.  Well, sort of: fail louder anyway.

Client quickly modified to pop-up an error dialog.  Its way to big to
fit on the screen so that needs improved in a future commit.