Abuse minimum deps to keep CI working on Ubuntu18.04
We don't really have a minimum version of WeasyPrint.  But on Ubuntu
18.04 we have a maximum version because of Issue #1594.  Our only CI job
that uses Ubuntu 18.04 is the "mindeps" job...  sooooooo... we pretend
our minimum verison of WeasyPrint is 52.5 and shazam that version is used
on Ubuntu 18.04.  This helps exactly no one except our CI runs.  And it
abuses the minimum deps feature as we don't REALLY need 52.5.