Ensure the underlying margin box cannot be deleted
This is a quick fix for #1750.  Probably we need a better fix like a
mixin class for objects which are UI and objects which are annotations.
That way we would not need these block lists of things that should not
be deleted...  But this seems to fix it for now!