Call same dialog from Annotator
This is nicer in a way b/c Annotator does not directly talk to
messengers.  And we can re-use the same dialog(s).  A downside is that
the parent of the dialog needs to be Annotator---otherwise the Marker
window pops up on top (at least on Gnome).  So I added a `parent` kwarg
which is then used to parent the dialogs.