New tempfile on each cat download
I was a bit uncomfortable with that tempfile being bother written and
read from given we don't understand the QPixmap that is reading it.
Seems safer to destroy it and build a new one on each download.  This
also works nicely  with exception handling because you can set `.catz`
to `None` and ImageViewWidget will DTRT.