Render to jpeg and keep that if smaller than png
On a real-world exam, rendered from larger scanned-produced poor-quality
jpegs-in-pdf, this gives a filesize reduction of about 43MiB -> 15MiB.
Some of that might be wasted effort on jpeg artifacts so the savings are
perhaps less significant with better scans.  But the code takes
whichever is better so shouldn't hurt (except it takes longer).