Better error in case of no vermap
The command line tool `plom-build` always tries to make the DB before
building papers, but if you're scripting, you can call `build_papers`
before `build_databaes` and this gives a cryptic "500 internal server
error" message without any indication of what is wrong.  5xx errors are
for server fault and perhaps this was a long time ago, but now if you
are building papers before populating the DB its because you (the
client) have made a mistake.  So use 409 Conflict instead.  404 was
also considered.

For the individual getQuestionVersionMap{papernum}, PlomRangeException
416 is a possibility but decided to wait for Issue #1745 before trying.