Marker: fix crash on question out of range
Fixes #1911.  Not so nice a fix: its hard to fail nicely during Marker
startup (which is probably not good).  I cleaned up the close code to
work if there is not yet a backgroupUploader, which helps.  #1911 is
regression b/c we used to get a somewhat crytic error message 416
question out of range.  Now it does that again; better than a crash.