PageScene: method to translate multiple items
Creates a single entry ("macro") on the undo stack.  Translates a set of
items on the page.  Input is a list of items; filter out anything that
isn't user-placed (which seems harder to do than it should be!)

Proof of concept: press Ctrl+Shift+H to translate roughly 50% of the
items 20 pixels leftward.