Don't push system rubrics when re-using rubrics
Delta rubrics and "no marks" / "full marks" rubrics are autogenered,
currently my Manager but perhaps should be HAL.  Anyway for now we can
just block on non-neutral non-relative as the others are system rubrics.

Leave a note in the code that this may need to be revisited as we
further refine the rubric system.  Fixes #1494 for now.