Explicit fail better than division by zero
If empty images or other junk got as far as the viewers, it would leader
to pix.height returning zero and then division by zero.  But this is not
a very clear message, and is just a symptom of something else.  So
explicitly fail if the pixmap `isNull` returns true.  Tested with inputs
like when the image data is `b"12"` (which I actually saw in some other
bug).