If marker gives back None, don't try to load it
Reset the scene before permuting: this is what saveAndGetNext does too.
This addreses some TODOs in the code, which ultimately led to this bug.
Change closeCurrentTGV to actually hide the scene: saveAnnotations was
doing this which doesn't make sense to me (and had a TODO to that
effect).  Only two places call closeCurrentTGV and we want both of them
to show no images in the case where Marker doesn't give us more/updated
data.  Fixes #1967.  Well, fixes part of #1967: part of that issue
remains (why permute needs to wait on any downloaders) but that's a fix
for another MR.