Some hardcoded timeouts made multiples of defaults
See Issue #1575 (this is not a fix).  I previously tried to temporarily
increase timeouts to 6 mins but because upload was hardcoded, it stayed
at 2 mins...  This change will prevent that in the future, but what we
really need is streaming uploads (?) or more generally, graceful
failures and handling of low-bandwidth and packet loss, ... and a pony.