use the API for loading spec
This was done a long time ago but I guess not for the background server
objects.  I redid specVerifier's demo method to share a bit more code
with some other method that wrote to files: they now share a common private
method for maniuplating the num_to_produce field.

Instead of using a class method to build the spec and demos etc we use
some object methods and just some raw code.