DB addrow: use only exceptions for failures
At least for this function, we abandon the `ok, status = ` form and use
exceptions.  This makes it easy to identify the "unexpected" errors
(here RuntimeErrors) from the  ones that are reasonably the caller's
fault.

Use this to refactor the return codes for the API call, which addresses
a TODO in the ManagerMessenger.

Docs docs docs.