Remove some old code, was breaking reset transform
Not sure what this was supposed to do...  But it interferes with
reseting the transform---in fact it seems designed to do that.  Looks
like it tried to maintain zoom etc between loading different images: not
sure why we'd want to do that as I'm trying to get the exact opposite.