Page Rearranger: don't warn on dupes if input had dupes
The point of this check is to ensure that Page Rearranger itself only
permutes stuff that was already in the input and does not duplicate
pages.  But if the input had dupes, which is can per Issue #2026 and
Issue #1573, then maybe that's bad or maybe its not but at anyrate its
the way things are with self-scanned work and its not Page Rearranger's
fault.

Fixes #2026.  Leaving #1573 open for the underlying issue of duplicated
uploads.