Fix unintended failures on spec/scan mismatch
The scan-with-tpv out of range code has probably never happened (?) as
The code called after would then fail (as the files being moved were
previously moved out of the way).

Maybe it *should* fail but but no in an accidental way.

Finally, isn't the lower end of range 1 not zero?