Use Windows-safe filename for client/server logs
Windows filenames cannot have colon.  Fixes #2127.  This is a real issue
for the client: crash on start-up.  For the server, maybe its more
theoretical, but maybe someone running the server using WSL or wants to
copy the log file off their unix machine to Windows.

For the client, we use the local timezone.  For the server I went with
`.utcnow()` to get "UTC" rather than "PDT".