Don't refresh progress bars widgets before parented
When doing init of the tab we build a lot of widgets before putting them
in a grid.  Problem is we also contact server, update them and force
visible.  Before they have a parent.  So Qts best guess is they are
top-level windows.  Here I instead update them all afterwards, in the
bigger init method.

Fixes #1184, a long standing bug.