Remove tightpage somehow fixes texfragment width
We had very wide images, being incorrectly cropped on the right.  For
some reason this fixes is on my Fedora 35 machine.  Not sure if it
breaks it on older latex installs however.

Fixes #2209.  At least it might.  At the very least we add a test that
seems to fail on Fedora 35 so that we can detect this.