Annotator: refactor shortcut storage in lists/dicts
We need to store these things so they are not GC.  But its unpleasant so
it must be the wrong pattern.  Maybe QAction is the right pattern.
Anyway "getattr"/"setattr" is even worse so instead store them in lists
and dicts, using a common naming theme, definitely underscored.

Also reformat the rubric shortcut keys with a lambda factory.