Return all predictions if predictor not specified
Previously, it was undefined which predictions
would be returned if predictor=None.
This kind of non-determinism is confusing;
it's more useful to return all predictions,
so that the dict can be iterated over, rather
than return some dict of arbitrary predictions can
potentially differ each time the method is invoked.

Note: there's a different return value structure due
to the allowance of multiple predictions per papernum.
If predictor is specified, there can be at most one
predictions per papernum, so no lists are introduced.