UI: indicate when predictions match, but likely wrong
With extra students who are not on the classlist, the predictors still
attempt to read and match the student_ids. In the case that LAP does
match Greedy, and Greedy has certainty less than 0.3, it's very likely
that both predictions tried to incorrectly associate the student with a
number on the classlist. Thus, flag with a red-ish colour to draw attention.