Update src_img_data files on demand
Avoid trying to update the stored src_img_data updates in callback.
It seems like multiple sources of truth: downloader knows best if we
have this image or not.  For example, the page cache might in the future
discard older images.

Instead, we have to update src_img_data before using it.  In practice we
were doing this anyway (checking for placeholders) so its not more
onerous.  Generalize it to a function that does that.

Also, take img_src_data from the plom_file data if we have some.