Annotator: don't zoom if no scene, shorter qtimer
I've said it before and I'll say it again: those QTimer hacks are bad
news.  Shorten them to 100ms.  Wait twice this long in the test.

Also when the zoom button is triggered, leave early if we have no scene.
This happens in tests b/c we do not load a scene.

Together this Fixes #2676.