Extract ID box from pushed ID page images
If run with the demo, this may fetch a smaller subset of images,
since it only gets Images that have been pushed from staging.

Kept the code for querying all Staging ID Images in scan_service.py for now,
although the command-line tool was moved to the Identify app and modified
to work with pushed images.

Fixes #2733