Added more info to filename on download
I did this to help profs teaching multiple courses who might want
additional information about when the data was captured. In the event
that the file is copied or sent somewhere else the time is still
saved in the name. Also --UTC so there isn't a space in the name.