Clean up the validate -> login path
Validating starts a messenger and if its good, we keep it the instance
variable for later to use, this means the messenger can track the state
e.g., of SSL so we don't have to shadow.

Remove the _legacy instance variable that track state and instead use the
messenger itself to track its own state.