sidToTest: managers or scanners
It used to be just for Scanners.  Client-side, we move to baseMessenger.
Server-side its still locked to manager or scanner (as before) but
previously only scanners could effectively call it.

Additionally, we return a third argument about whether it was ID'd or
prenamed.  Technically an API break but no one checks.  Messenger deals
with both cases.