Fix the regexp for blank or whitespace only strings
Comments linking these two interrelated bits of code together.  I'm not
entirely happy with this: but short of writing my own QValidator
subclass I'm not sure how to harmonize the two approaches.  In some
sense, we want "anything that can be converted to a list of ints" and we
have two implementations of that...  Not ideal.