Try harder to shorten abbreviated text in catch-all dialog
The chain of exceptions is often pretty long (esp. b/c this is the
unexpected error handler).  The "trim to 8 lines" was apparently not
doing its job, perhaps b/c the docs mention that there can be addition
newlines inside each element of this list.

We also shorten each line to 80 chars.

The full raw output is still available under details as before.

Fixes #3217.