Parmeterized rubrics are no longer experimental
This was disabled because of #2462.  Well, we can now lockdown who can
create/modify rubrics server-side.  We don't yet have the granularity to
restrict this specific feature and I'm not sure it would be worth the
effort.

More importantly, this has been used in practice and the only feedback
was that users forgot to re-enable experimental mode after a
client-restart.  Ergo, the experimental label was what got in the way,
rather than this feature itself.