client: check server API version on start
The server used to do all this for us, but we relaxed to the server only
caring if the client is *new enough*.  It does not check if the client
is *too new*.  Now the client checks if the server is too old for it to
support.

This will allow more flexible client-server support where the client can
continue to work with older servers, for essentially as long as the
client's `Messenger` supports older API versions.

Fixes #3247.