move rubric change reaction to checklegality
This means it gets called on Annotator open, in fact a little too
aggressively as there is no included rubric list.  In this case, each
return as there is nothing we can check.

But leave the debugging code in just in case we cannot find a rubric
that we expect too (which would otherwise crash).